# Documents/ApiSpec/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "API",
    "description": "API Spec f�r den digitalen Reisef�hrer",
    "version": "v1"
  },
  "paths": {
...
    "/api/route": {
      "post": {
        "tags": [
          "Route"
        ],
        "summary": "Add a new route to the database",
        "operationId": "AddRoute",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PostRoute"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostRoute"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostRoute"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostRoute"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Route"
        ],
        "summary": "Edits the route to a given id",
        "operationId": "PutRoute",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PutRoute"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutRoute"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PutRoute"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PutRoute"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/route/{routeID}": {
      "delete": {
        "tags": [
          "Route"
        ],
        "summary": "Deletes the route to a given id",
        "operationId": "DeleteRoute",
        "parameters": [
          {
            "name": "routeID",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Route"
        ],
        "summary": "Gets the route to a given id",
        "operationId": "GetRoute",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "routeID",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/routes/{userName}": {
      "get": {
        "tags": [
          "Route"
        ],
        "summary": "Get all routes from the given user",
        "operationId": "GetRoutes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Route"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/routes/all": {
      "get": {
        "tags": [
          "Route"
        ],
        "summary": "Get all routes",
        "operationId": "GetAllRoutes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Route"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PointOfInterest": {
        "required": [
          "averageRating",
          "description",
          "imageUrl",
          "isPromoted",
          "latitude",
          "longitude",
          "name",
          "numberOfRatings",
          "poIID",
          "userName"
        ],
        "type": "object",
        "properties": {
          "poIID": {
            "type": "string",
            "description": "Gets or Sets Id",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gets or Sets Name"
          },
          "userName": {
            "type": "string",
            "description": "Gets or Sets username"
          },
          "description": {
            "type": "string",
            "description": "Gets or Sets Description"
          },
          "latitude": {
            "type": "number",
            "description": "Gets or Sets Latitude",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "description": "Gets or Sets Longitude",
            "format": "double"
          },
          "averageRating": {
            "type": "number",
            "description": "Gets or Sets AverageRating",
            "format": "double"
          },
          "numberOfRatings": {
            "type": "integer",
            "description": "Gets or Sets NumberOfRatings",
            "format": "int32"
          },
          "imageUrl": {
            "type": "string",
            "description": "Gets or Sets ImageUrl"
          },
          "isPromoted": {
            "type": "boolean",
            "description": "Defines, whether this poi is promoted or not"
          }
        },
        "additionalProperties": false,
        "description": "Model class for a PointOfInterest"
      },
...
      "PostRoute": {
        "required": [
          "creatorName",
          "description",
          "duration",
          "name",
          "pointOfInterests",
          "polyline"
        ],
        "type": "object",
        "properties": {
          "pointOfInterests": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Gets or Sets PointOfInterests"
          },
          "name": {
            "type": "string",
            "description": "Gets or Sets Name"
          },
          "description": {
            "type": "string",
            "description": "Gets or Sets Description"
          },
          "creatorName": {
            "type": "string",
            "description": "Gets or Sets CreatorName"
          },
          "duration": {
            "type": "number",
            "description": "Gets or Sets Duration",
            "format": "float"
          },
          "polyline": {
            "type": "string",
            "description": "Gets or Sets Polyline"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "PutRoute": {
        "required": [
          "creatorName",
          "description",
          "duration",
          "id",
          "name",
          "pointOfInterests",
          "polyline"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or Sets Id"
          },
          "pointOfInterests": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Gets or Sets PointOfInterests"
          },
          "name": {
            "type": "string",
            "description": "Gets or Sets Name"
          },
          "description": {
            "type": "string",
            "description": "Gets or Sets Description"
          },
          "creatorName": {
            "type": "string",
            "description": "Gets or Sets CreatorName"
          },
          "duration": {
            "type": "number",
            "description": "Gets or Sets Duration",
            "format": "float"
          },
          "polyline": {
            "type": "string",
            "description": "Gets or Sets Polyline"
          }
        },
        "additionalProperties": false
      },
...
      "Route": {
        "required": [
          "averageRating",
          "creatorName",
          "description",
          "duration",
          "name",
          "numberOfRatings",
          "polyline",
          "routeId"
        ],
        "type": "object",
        "properties": {
          "routeId": {
            "type": "string",
            "description": "Gets or Sets Id",
            "format": "uuid"
          },
          "pointOfInterests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PointOfInterest"
            },
            "description": "The PoIs that are on this route",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Gets or Sets Name"
          },
          "description": {
            "type": "string",
            "description": "Gets or Sets Description"
          },
          "creatorName": {
            "type": "string",
            "description": "Gets or Sets CreatorName"
          },
          "duration": {
            "type": "number",
            "description": "Gets or Sets Duration",
            "format": "float"
          },
          "polyline": {
            "type": "string",
            "description": "Gets or Sets Polyline"
          },
          "averageRating": {
            "type": "number",
            "description": "Gets or Sets AverageRating",
            "format": "double"
          },
          "numberOfRatings": {
            "type": "integer",
            "description": "Gets or Sets NumberOfRatings",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}